+++ include/objects/jumper.h
@@ -5,9 +5,9 @@
 
 class TA_Jumper : public TA_Object {
 private:
-    const double jumpSpeed = 0.3;
-    const double gravity = 0.085;
-    const double idleTime = 40;
+    const float jumpSpeed = 0.3;
+    const float gravity = 0.085;
+    const float idleTime = 40;
 
     enum State { STATE_IDLE, STATE_AIM, STATE_JUMP };
 
@@ -26,7 +26,7 @@ private:
     State state = STATE_IDLE;
     TA_Point velocity;
     bool direction = false;
-    double timer = 0;
+    float timer = 0;
 
 public:
     using TA_Object::TA_Object;
